backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
tortoise-orm[asyncpg]>=0.20.0
aerich>=0.7.2
python-multipart
typer>=0.9.0
httpx>=0.26.0
email-validator
passlib[bcrypt]
python-jose[cryptography]
